+++ Assets/Scripts/Notes/NotePosition.cs
@@ -42,7 +42,7 @@ namespace NoteMaker.Notes
         /// この位置をサンプル位置に変換します。
         /// BPM と LPB に基づいて、拍位置 → サンプル数へ変換します。
         /// </summary>
-        public int ToSamples(int frequency, int BPM, System.Collections.Generic.IEnumerable<NoteMaker.Model.BpmChange> changes = null)
+        public int ToSamples(int frequency, float BPM, System.Collections.Generic.IEnumerable<NoteMaker.Model.BpmChange> changes = null)
         {
             return NoteMaker.Utility.BPMUtility.CalculateSamples(frequency, BPM, LPB, num, changes);
         }
